move ProcessReader into Lib and rename it to PhpProcessReader
