+++ src/Command/Inspector/GetTraceCommand.php
@@ -16,9 +16,9 @@ namespace PhpProfiler\Command\Inspector;
 use PhpProfiler\Lib\Elf\Parser\ElfParserException;
 use PhpProfiler\Lib\Elf\Process\ProcessSymbolReaderException;
 use PhpProfiler\Lib\Elf\Tls\TlsFinderException;
+use PhpProfiler\Lib\PhpProcessReader\PhpGlobalsFinder;
 use PhpProfiler\Lib\Process\MemoryReader\MemoryReaderException;
-use PhpProfiler\ProcessReader\PhpGlobalsFinder;
-use PhpProfiler\ProcessReader\PhpMemoryReader\ExecutorGlobalsReader;
+use PhpProfiler\Lib\PhpProcessReader\PhpMemoryReader\ExecutorGlobalsReader;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
